update card and move step indicator to common components
