+++ frontend/src/app/commonComponents/CardContainer.tsx
@@ -13,8 +13,9 @@ export const CardContainer: React.FC<CardContainerProps> = ({
       <div className="grid-container maxw-tablet usa-section">
         <div className="card width-mobile-lg">
           {logo && (
-            <header>
-              <img src={siteLogo} alt="SimpleReport" />
+            <header className="display-flex flex-column">
+              <img className="maxw-card-lg flex-align-self-center" src={siteLogo} alt="SimpleReport" />
+              <div className="border-bottom border-base-lighter margin-x-neg-3 margin-top-3"></div>
             </header>
           )}
           {children}
