+++ frontend/src/app/commonComponents/StepIndicator.tsx
